API/requirements.txt
python-dotenv==1.0.0
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic[email]
PyJWT==2.6.0
python-decouple==3.7
mysql-connector-python==8.0.31